requirements.txt
pytest==7.4.2
selenium==4.13.0
allure-pytest==2.13.2
# urllib3==1.26.6
discord-webhook==1.3.0
python-dotenv==1.0.0